Move xcodebuild and iproxy into their own modules
