+++ lib/driver.js
@@ -2,7 +2,7 @@ import { BaseDriver } from 'appium-base-driver';
 import { util } from 'appium-support';
 import _ from 'lodash';
 import { launch } from 'node-simctl';
-import WebDriverAgent from './webdriveragent';
+import WebDriverAgent from './wda/webdriveragent';
 import log from './logger';
 import { simBooted, createSim, getExistingSim, runSimulatorReset,
          installToSimulator } from './simulator-management';
