+++ lib/utils.js
@@ -1,4 +1,4 @@
-import { fs, tempDir } from 'appium-support';
+import { fs } from 'appium-support';
 import path from 'path';
 import { utils as iosUtils } from 'appium-ios-driver';
 import { exec } from 'teen_process';
@@ -77,15 +77,6 @@ async function getAndCheckIosSdkVersion () {
   return versionNumber;
 }
 
-async function killAppUsingAppName (udid, appName) {
-  let psArgs = [`-c`, `ps -ax|grep -i "${appName}"|grep -i "${udid}"|grep -v grep|awk '{print "kill -9 " $1}'|sh`];
-  try {
-    await exec(`bash`, psArgs);
-  } catch (err) {
-    log.debug(`Error : ${err.message}`);
-  }
-}
-
 async function adjustWDAAttachmentsPermissions (perms) {
   if (!process.env.HOME) {
     throw new Error('Need HOME env var to be set in order to adjust WDA attachments permission');
@@ -118,18 +109,6 @@ async function clearSystemFiles (wda) {
   await iosUtils.clearLogs(toDelete);
 }
 
-async function generateXcodeConfigFile (orgId, signingId) {
-  log.debug(`Generating xcode config file for orgId '${orgId}' and signingId ` +
-            `'${signingId}'`);
-  let contents = `DEVELOPMENT_TEAM = ${orgId}
-CODE_SIGN_IDENTITY = ${signingId}
-`;
-  let xcconfigPath = await tempDir.path('appium-temp.xcconfig');
-  log.debug(`Writing xcode config file to ${xcconfigPath}`);
-  await fs.writeFile(xcconfigPath, contents, "utf8");
-  return xcconfigPath;
-}
-
 async function checkAppPresent (app) {
   log.debug(`Checking whether app '${app}' is actually present on file system`);
   if (!(await fs.exists(app))) {
@@ -149,6 +128,5 @@ async function getDriverInfo () {
 }
 
 export { detectUdid, getAndCheckXcodeVersion, getAndCheckIosSdkVersion,
-         killAppUsingAppName, adjustWDAAttachmentsPermissions,
-         generateXcodeConfigFile, checkAppPresent, getDriverInfo,
+         adjustWDAAttachmentsPermissions, checkAppPresent, getDriverInfo,
          clearSystemFiles };
