+++ lib/wda/no-session-proxy.js
@@ -1,6 +1,7 @@
 import { util } from 'appium-support';
 import { JWProxy } from 'appium-base-driver';
 
+
 class NoSessionProxy extends JWProxy {
   constructor (opts = {}) {
     super(opts);
