+++ lib/wda/utils.js
@@ -1,7 +1,7 @@
-import { fs } from 'appium-support';
+import { fs, tempDir } from 'appium-support';
 import { exec } from 'teen_process';
 import path from 'path';
-import log from './logger';
+import log from '../logger';
 
 
 const WDA_RUNNER_BUNDLE_ID = 'com.facebook.WebDriverAgentRunner';
@@ -109,5 +109,51 @@ async function fixForXcode7 (bootstrapPath, initial = true) {
   await fixXCUICoordinateFile(bootstrapPath, initial);
 }
 
+async function generateXcodeConfigFile (orgId, signingId) {
+  log.debug(`Generating xcode config file for orgId '${orgId}' and signingId ` +
+            `'${signingId}'`);
+  let contents = `DEVELOPMENT_TEAM = ${orgId}
+CODE_SIGN_IDENTITY = ${signingId}
+`;
+  let xcconfigPath = await tempDir.path('appium-temp.xcconfig');
+  log.debug(`Writing xcode config file to ${xcconfigPath}`);
+  await fs.writeFile(xcconfigPath, contents, "utf8");
+  return xcconfigPath;
+}
+
+async function killAppUsingAppName (udid, appName) {
+  let psArgs = [`-c`, `ps -ax|grep -i "${appName}"|grep -i "${udid}"|grep -v grep|awk '{print "kill -9 " $1}'|sh`];
+  try {
+    await exec(`bash`, psArgs);
+  } catch (err) {
+    log.debug(`Error : ${err.message}`);
+  }
+}
+
+async function killProcess (name, proc) {
+  if (proc && proc.proc) {
+    log.info(`Shutting down ${name} process (pid ${proc.proc.pid})`);
+    try {
+      await proc.stop('SIGTERM', 1000);
+    } catch (err) {
+      if (err.message.indexOf(`Process didn't end after`) === -1) {
+        throw err;
+      }
+      log.debug(`${name} process did not end in a timely fashion: '${err.message}'. ` +
+                `Sending 'SIGKILL'...`);
+      try {
+        await proc.stop('SIGKILL');
+      } catch (err) {
+        if (err.message.indexOf('not currently running') !== -1) {
+          // the process ended but for some reason we were not informed
+          return;
+        }
+        throw err;
+      }
+    }
+  }
+}
+
 export { updateProjectFile, resetProjectFile, checkForDependencies,
-         setRealDeviceSecurity, fixForXcode7 };
+         setRealDeviceSecurity, fixForXcode7, generateXcodeConfigFile,
+         killAppUsingAppName, killProcess };
